+++ jsonschema-module-swagger-2/src/test/java/com/github/victools/jsonschema/module/swagger2/IntegrationTest.java
@@ -97,7 +97,7 @@ public class IntegrationTest {
         @Schema(description = "field description", nullable = true, allowableValues = {"A", "B", "C", "D"}, minLength = 1, maxLength = 1)
         public String fieldWithDescriptionAndAllowableValues;
 
-        @Schema(minimum = "15", maximum = "20", multipleOf = 0.01234567891)
+        @Schema(minimum = "15", maximum = "20", multipleOf = 0.0123456789)
         public BigDecimal fieldWithInclusiveNumericRange;
 
         @Schema(minimum = "14", maximum = "21", exclusiveMinimum = true, exclusiveMaximum = true, required = true, multipleOf = 0.1)